// src/types/movie.ts
export interface MovieDetail {
  vod_id: string;
  vod_name: string;
  vod_pic: string;
  vod_remarks: string;
  vod_year: string;
  vod_score: string;
  vod_content: string;
  vod_play_url: string;
  vod_actor: string;
  vod_director: string;
  vod_area: string;
  vod_lang: string;
}

export interface PlayUrl {
  name: string;
  url: string;
}